+++ examples/integration_test_example/lib/main.dart
@@ -1,0 +1,5 @@
+import 'package:widgetbook/widgetbook.dart';
+
+import 'widgetbook.config.dart';
+
+void main() => runWidgetbook(config);